requirements.txt
Flask>=2.3.0
SQLAlchemy>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.15
openai>=1.3.0
lightrag-hku>=0.1.5
httpx>=0.27
pydantic>=2.7
python-dotenv>=1.0